1.19.4 part 4

+++ src/main/java/com/github/alexthe666/alexsmobs/entity/ai/ElephantAIForageLeaves.java
@@ -78,7 +78,7 @@ public class ElephantAIForageLeaves extends MoveToBlockGoal {
     }
 
     private boolean isWithinXZDist(BlockPos blockpos, Vec3 positionVec, double distance) {
-        return blockpos.distSqr(new BlockPos(positionVec.x(), blockpos.getY(), positionVec.z())) < distance * distance;
+        return blockpos.distSqr(new BlockPos((int) positionVec.x(), blockpos.getY(), (int) positionVec.z())) < distance * distance;
     }
 
     protected boolean isReachedTarget() {
