+++ src/main/java/com/github/alexthe666/alexsmobs/entity/ai/EndergradeAIBreakFlowers.java
@@ -63,7 +63,7 @@ public class EndergradeAIBreakFlowers extends MoveToBlockGoal {
     }
 
     private boolean isWithinXZDist(BlockPos blockpos, Vec3 positionVec, double distance) {
-        return blockpos.distSqr(new BlockPos(positionVec.x(), blockpos.getY(), positionVec.z())) < distance * distance;
+        return blockpos.distSqr(new BlockPos((int) positionVec.x(), blockpos.getY(), (int) positionVec.z())) < distance * distance;
     }
 
     protected boolean isReachedTarget() {
