+++ src/main/java/com/github/alexthe666/alexsmobs/entity/ai/LeafcutterAntAIForageLeaves.java
@@ -92,7 +92,7 @@ public class LeafcutterAntAIForageLeaves extends MoveToBlockGoal {
                 if (Math.abs(xDif) < 0.6 && Math.abs(zDif) < 0.6) {
                     ant.setDeltaMovement(ant.getDeltaMovement().multiply(0D, 1D, 0D));
                     this.ant.getMoveControl().setWantedPosition(logStartPos.getX() + 0.5D, ant.getY() + 2, logStartPos.getZ() + 0.5D, 1);
-                    BlockPos test = new BlockPos(logStartPos.getX(), ant.getY(), logStartPos.getZ());
+                    BlockPos test = new BlockPos(logStartPos.getX(), (int) ant.getY(), logStartPos.getZ());
                     if (!ant.level.getBlockState(test).is(BlockTags.LOGS) && ant.getAttachmentFacing() == Direction.DOWN) {
                         this.stop();
                         return;
